+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Schemas to components to be able to edit them via Site Editor.
+- `condition.product` block.
+- `enabled` flag for conditions in the site editor.
 
 ## [1.0.1] - 2020-04-08
 ### Fixed